fix: removes redux and place sign up and log in functionality in components
