--- store/index.js
@@ -1,14 +1,0 @@
-import { combineReducers, createStore, applyMiddleware } from 'redux';
-import thunkMiddleware from 'redux-thunk';
-import { composeWithDevTools } from 'redux-devtools-extension';
-import { createLogger } from 'redux-logger';
-import user from './user';
-
-const rootReducer = combineReducers({
-  user,
-});
-
-const middleWare = composeWithDevTools(
-  applyMiddleware(thunkMiddleware, createLogger({ collapsed: true }))
-);
-export const store = createStore(rootReducer, middleWare);
